Improve IStrategy and AbstractStrategy
- Expose getAssetAmount() and adjust()
- Add sweep() to AbstractStrategy for retrieving unrelated tokens
- Misc improvements

+++ contracts/strategies/alphahomora/interfaces/IYearnToken.sol
@@ -6,7 +6,8 @@ pragma solidity 0.8.6;
 // note it's actually just using compound contract so we could use full CTokenInterfaces.sol
 interface IYearnToken {
     // exchange rate scaled by 1e18
-    function exchangeRateStored() external view returns (uint);
+    function exchangeRateStored() external view returns (uint256);
+
     // current to ensure token computed for sell is accurate
-    function exchangeRateCurrent() external returns (uint);
+    function exchangeRateCurrent() external returns (uint256);
 }
